Adding existingUserID constant to Story default
Successful Story creation requires the user of a UserID that is stored in the database. In order to not depend on the Login team (while dev'ing) I added a existing UserID. This ID is never changed and any story that is created via our tool will be under this user.