Update backend URL (#89)
* Revert "added .env to gitignore (#68)"

This reverts commit 64945c56b1f3a597dcfb2b8a8bf9a1d03b60c4e5.

* .env file: add API environment variable

We will use this to allow changing the API the front-end points its API
requests to.

* API: point calls to API env var

* README: add token instructions

These instructions should be followed to add the Mapbox token as an
enviroment variable. We needed to move this to `.env.local` so that it
would not be tracked when adding `.env` to source control.

* API env var: add /

This fixes an error that was causing enpoints to be appended to the url
without a `/`. i.e. `.comtoken` rather than `.com/token`.

* CORS: don't set credentials flag

Our server is configured with `Access-Control-Allow-Origin = *` for the
time being. We would like to change this when we have a URL. For the
time being we need to stop setting the credentials header on the
front-end requests. See
https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS/Errors/CORSNotSupportingCredentials.