Login: make form (#98)
* Login: make form

Making this a form makes intuitive sense. More importantly, it makes it
easy for the user to login by hitting the "enter" key in addition to
clicking the button.

* Form: prevent default

Forms by default attempt to submit input data to the server. We don't
need that since we are specifying our on behavior for `onSubmit`.
Therefore, we need to `preventDefault`.

* Eslint: fix formatting errors