Test api base (#64)
* constructGetParameters: fix type

* constructGetParameters: add tests

This tests three different scenerios for testing the
`constructGetParameters` method. The first is a single key value pair,
the second two, and the third none. This method is supposed to construct
get parameters using the "? and &" notation based on key value pairs and
return a string to be appended to a url.

* Lint: run Prettier