Remove admin shortcut buttons on story card (#299)
This PR removes the "Reject" and "Publish" shortcut buttons that used to appear on story cards. To be clear, removing the buttons does not eliminate the functionality. Admins can still click on the story card to view the full-length story, and the review UI to approve or reject with feedback will appear at the bottom. These buttons were intended as a convenience, but connecting them properly is more trouble than it's worth.

* removed story card admin shortcuts

* removed redundant import